requirements.txt
pyTelegramBotAPI==4.14.0
google-generativeai==0.3.2
feedparser==6.0.11
schedule==1.2.1
streamlit==1.32.0
pandas==2.2.1